refactor(userscript): improve userscript API handling and fallback strategy refactor(events): update DOMEventManager to use EventManager for listener management refactor(zip): enforce ZIP format limits in StreamingZipWriter refactor(download): enhance filename handling in downloadAsZip function refactor(notification): simplify notification service to use safe userscript API refactor(persistent-storage): streamline persistent storage to use safe userscript API refactor(retry): refine retry logic and result types for better clarity

+++ src/shared/network/retry-fetch.ts
@@ -45,7 +45,7 @@ export async function fetchArrayBufferWithRetry(
   );
 
   if (result.success) {
-    return result.data as Uint8Array;
+    return result.data;
   }
 
   // If the caller's signal is aborted (including during retry backoff),
